+++ src/main/java/chess/domain/piece/King.java
@@ -1,10 +1,16 @@
-package chess;
+package chess.domain.piece;
 
-public class Knight extends Piece {
+import chess.domain.position.Position;
+import chess.domain.piece.info.Team;
 
-    public Knight(Team team) {
+public class King extends Piece {
+
+    private static final int UPPER_BOUND_OF_MOVABLE_DISTANCE = 1;
+
+    public King(Team team) {
         super(team);
     }
+
 
     @Override
     public boolean canMove(final Position startPosition, final Position endPosition) {
@@ -14,7 +20,8 @@ public class Knight extends Piece {
         int diffFile = endPosition.calculateFileDistance(startPosition);
         int diffRank = endPosition.calculateRankDistance(startPosition);
 
-        return Math.abs(diffFile * diffRank) == 2;
+        return Math.abs(diffFile) <= UPPER_BOUND_OF_MOVABLE_DISTANCE
+                && Math.abs(diffRank) <= UPPER_BOUND_OF_MOVABLE_DISTANCE;
     }
 
     @Override
