+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,8 +1,11 @@
-package chess;
+package chess.domain.piece;
 
-public class Rook extends Piece {
+import chess.domain.position.Position;
+import chess.domain.piece.info.Team;
 
-    public Rook(final Team team) {
+public class Queen extends Piece {
+
+    public Queen(final Team team) {
         super(team);
     }
 
@@ -14,7 +17,7 @@ public class Rook extends Piece {
         int diffFile = endPosition.calculateFileDistance(startPosition);
         int diffRank = endPosition.calculateRankDistance(startPosition);
 
-        return (diffFile * diffRank == 0);
+        return (Math.abs(diffFile) == Math.abs(diffRank)) || (diffFile * diffRank == 0);
     }
 
     @Override
