+++ src/main/java/chess/domain/piece/info/Log.java
@@ -1,4 +1,6 @@
-package chess;
+package chess.domain.piece.info;
+
+import chess.domain.position.Position;
 
 public class Log {
