+++ src/main/java/chess/domain/piece/info/Team.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain.piece.info;
 
 public enum Team {
     BLACK,
