+++ src/test/java/chess/domain/piece/info/TraceTest.java
@@ -1,7 +1,11 @@
-package chess;
+package chess.domain.piece.info;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.piece.info.Trace;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
