+++ src/test/java/chess/domain/position/PositionTest.java
@@ -1,7 +1,8 @@
-package chess;
+package chess.domain.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.piece.info.Team;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
